+++ src/Plugin/ParallelExecution/Event/ExecuteEnd.php
@@ -13,8 +13,6 @@ use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * ExecuteEnd event, see Events for details.
- *
- * @codeCoverageIgnore
  */
 class ExecuteEnd extends Event
 {